create training_type subpackage
