+++ pytorch_lightning/accelerators/plugins/__init__.py
@@ -1,0 +1,3 @@
+from pytorch_lightning.accelerators.plugins.base_plugin import Plugin
+from pytorch_lightning.accelerators.plugins.precision import *
+from pytorch_lightning.accelerators.plugins.training_type import *
