+++ pytorch_lightning/accelerators/plugins/training_type/__init__.py
@@ -1,0 +1,8 @@
+from pytorch_lightning.accelerators.plugins.training_type.ddp import DDPPlugin
+from pytorch_lightning.accelerators.plugins.training_type.ddp2 import DDP2Plugin
+from pytorch_lightning.accelerators.plugins.training_type.ddp_spawn import DDPSpawnPlugin
+from pytorch_lightning.accelerators.plugins.training_type.dp import DataParallelPlugin
+from pytorch_lightning.accelerators.plugins.training_type.horovod import HorovodPlugin
+from pytorch_lightning.accelerators.plugins.training_type.parallel import ParallelPlugin
+from pytorch_lightning.accelerators.plugins.training_type.single_device import SingleDevicePlugin
+from pytorch_lightning.accelerators.plugins.training_type.training_type_plugin import TrainingTypePlugin
