+++ pytorch_lightning/accelerators/plugins/training_type/ddp2.py
@@ -1,0 +1,5 @@
+from pytorch_lightning.accelerators.plugins.training_type.ddp import DDPPlugin
+
+# TODO: DDP2
+class DDP2Plugin(DDPPlugin):
+    pass
