+++ pytorch_lightning/accelerators/plugins/training_type/single_device.py
@@ -1,0 +1,40 @@
+import torch
+from pytorch_lightning.accelerators.plugins.training_type.training_type_plugin import TrainingTypePlugin
+
+
+class SingleDevicePlugin(TrainingTypePlugin):
+    def __init__(self, device):
+        super().__init__()
+        self.device: torch.device = device
+
+    @property
+    def on_gpu(self):
+        return self.device.type == "cuda" and torch.cuda.is_available()
+
+    def reduce(self, output, *args, **kwargs):
+        return output
+
+    @property
+    def root_device(self):
+        return self.device
+    
+    def model_to_device(self):
+        if self.on_gpu:
+            torch.cuda.set_device(self.root_device)
+
+        self._model.to(self.root_device)
+
+    def connect(self, model: torch.nn.Module):
+        self._model = model
+        self.model_to_device()
+        return self.model
+
+    @property
+    def is_global_zero(self):
+        return True
+
+    def barrier(self, *args, **kwargs):
+        pass
+
+    def broadcast(self, obj: object, src: int = 0) -> object:
+        return obj
